Moved process_optics_geometry to app/sofast/lib
