+++ opencsp/app/sofast/lib/ProcessSofastFixed.py
@@ -15,7 +15,7 @@ from opencsp.common.lib.deflectometry.BlobIndex import BlobIndex
 import opencsp.common.lib.deflectometry.calculation_data_classes as cdc
 from opencsp.common.lib.deflectometry.FacetData import FacetData
 import opencsp.common.lib.deflectometry.image_processing as ip
-import opencsp.common.lib.deflectometry.process_optics_geometry as pr
+import opencsp.app.sofast.lib.process_optics_geometry as pr
 from opencsp.common.lib.deflectometry.SlopeSolver import SlopeSolver
 from opencsp.common.lib.deflectometry.SlopeSolverData import SlopeSolverData
 from opencsp.app.sofast.lib.SpatialOrientation import SpatialOrientation
